Create README.md
These projects were completed as part of a course teaching low-level computing concepts. The goal is to build a computer platform from the ground up, from hardware all the way to finally developing an operating system to run on the program.

The project uses a group of domain-specific languages, including an HDL, an assembly language called Hack, and a high-level object-oriented language based on Java, called Jack. 

For the software portion, I've also chosen to implement the computer's assembler, virtual machine, and compiler using JavaScript.